% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..bases..}
\alias{BASE$..bases..}
\title{Direct parent classes}
\description{
Direct parent classes
}
\examples{

BASE$..bases..

# Inherit from BASE
TEST <- new_class(BASE, class_name = "TEST")
TEST$..bases..
}
