% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..class..}
\alias{BASE$..class..}
\title{Class name and parent class names}
\description{
A string vector.
}
\examples{

BASE$..class..

# Inherit from BASE
TEST <- new_class(BASE, class_name = "TEST")
TEST$..class..
}
