% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..class_tree..}
\alias{BASE$..class_tree..}
\title{Class name and parent class names represented in a tree}
\description{
A list.
}
\examples{

BASE$..class_tree..

# Inherit from BASE
TEST <- new_class(BASE, class_name = "TEST")
TEST$..class_tree..
}
