% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..dict..}
\alias{BASE$..dict..}
\title{All names in the class or instance environment}
\value{
A vector of string.
}
\description{
This function returns all names in the environment.
}
\examples{

BASE$..dict..()

# Instantiate
test <- BASE$instantiate()
test$..dict..()
}
