% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..dir..}
\alias{BASE$..dir..}
\title{All names in the class or instance environment}
\value{
A vector of string.
}
\description{
This function returns all names in the environment.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{BASE$..dir..()
}\if{html}{\out{</div>}}
}
}
\examples{

BASE$..dir..()

# Instantiate
test <- BASE$instantiate()
test$..dir..()
}
