% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..doc_pkg..}
\alias{BASE$..doc_pkg..}
\title{Class documentation location}
\description{
Name of the package providing the class documentation,
referenced as \verb{?\{doc_pkg\}::\{class\}}, where \code{{class}} is the class name.
}
\examples{

BASE$..doc_pkg..

DEMO <- new_class(BASE, class_name = "DEMO", doc_pkg = "test")
DEMO$..doc_pkg..
}
