% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..instantiated..}
\alias{BASE$..instantiated..}
\title{Instantiate status}
\description{
Whether or not the object is an instance.
}
\examples{

BASE$..instantiated..

# Instantiate
test <- BASE$instantiate()
test$..instantiated..
}
