% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..method_env..}
\alias{BASE$..method_env..}
\title{The container}
\description{
A container where methods will be executed.
}
\examples{

BASE$..method_env..

# Inherit from BASE
TEST <- new_class(BASE, class_name = "TEST")
TEST$..method_env..
}
