% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..mro..}
\alias{BASE$..mro..}
\title{Method resolution order}
\description{
Method resolution order defined using C3 algorithm.
}
\examples{

BASE$..mro..

# Inherit from BASE
TEST <- new_class(BASE, class_name = "TEST")
TEST$..mro..
}
