% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..repr..}
\alias{BASE$..repr..}
\title{"Official" String representation of the object}
\value{
A string.
}
\description{
This function returns a "official" string representation of
the object, which may be used to reconstruct the object given an appropriate
environment.
}
\examples{

BASE$..repr..()

test <- base_()
test$..repr..()

test <- BASE$instantiate()
test$..repr..()

test <- BASE$..new..()
test$..repr..()
}
