% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..str..}
\alias{BASE$..str..}
\title{String representation of the object}
\value{
A string.
}
\description{
This function returns a string representation of the object.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{BASE$..str..()
}\if{html}{\out{</div>}}
}
}
\examples{

BASE$..str..()

# Inherit from BASE
TEST <- new_class(BASE, class_name = "TEST")
TEST$..str..()

# Instantiate
test <- BASE$instantiate()
test$..str..()
}
