% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$get_attr}
\alias{BASE$get_attr}
\title{Get value of an attribute or a method}
\arguments{
\item{attr_name}{Character. Attribute name.}
}
\value{
The attribute value.
}
\description{
This function gets the value of an attribute or a method.
}
\examples{

BASE$get_attr("test")

BASE$get_attr("..methods..")
}
