% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$has_attr}
\alias{BASE$has_attr}
\title{Whether or not an attribute or method exists}
\arguments{
\item{attr_name}{Character. Attribute name.}
}
\value{
True or FALSE.
}
\description{
This function checks whether or not an attribute or method
exists.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{BASE$has_attr(attr_name)
}\if{html}{\out{</div>}}
}
}
\examples{

BASE$has_attr("test")

BASE$has_attr("..len..")
}
