% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$instantiate}
\alias{BASE$instantiate}
\title{Instantiate method}
\arguments{
\item{...}{Arguments passed to \code{..init..} method.}

\item{env}{Environment. The instance environment.}

\item{init_call}{Call. Contents of the \code{..init_call..}. It is recommended to
leave it as default. Default is \code{init_call = sys.call()}.}
}
\value{
An instance environment.
}
\description{
This function will new an instance using the \code{..new..} method,
then initialized the instance with the \code{..init..} method.
}
\examples{

BASE$..dir..()

# Build an instance
base_instance <- BASE$instantiate()

base_instance$..dir..()
}
