% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM$..str..}
\alias{CLOSED_FORM$..str..}
\title{String representation of the object}
\value{
A string.
}
\description{
This function returns a string representation of the object.
}
\examples{

CLOSED_FORM$..str..()

test <- closed_form(~1+1)
test$..str..()
}
