% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM$as_dataframe}
\alias{CLOSED_FORM$as_dataframe}
\title{Transforming list to data frame}
\arguments{
\item{dat}{Vector or List. Results generated by \code{gen} method.}

\item{lhs}{Character. The name of the expression. Default is ".lhs".}
}
\value{
A data frame.
}
\description{
This function transforms the result generated by \code{gen} to a
data frame.
}
\examples{

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate data
result1 <- cf$gen(10)

cf$as_dataframe(result1)

# Generate data and keep the RHS
result2 <- cf$gen(10, rhs_val = TRUE)

# Specify the name
cf$as_dataframe(result2, lhs = "y")
}
