% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM$compute}
\alias{CLOSED_FORM$compute}
\title{Compute the expression without generating any random values}
\value{
Numeric value.
}
\description{
This function computes a deterministic expression.
}
\examples{

test <- closed_form(~1+mean(c(1,2,3)))
test$compute()
}
