% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM$expr}
\alias{CLOSED_FORM$expr}
\title{Expression extracted from the provided formula}
\description{
A language object.
}
\examples{

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(omit~omit~3 * (exp(a) + b))

# Only the right hand side of the last `~` will be kept
cf$expr

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

cf2$expr
}
