% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM$set_sym}
\alias{CLOSED_FORM$set_sym}
\title{Set values for symbols}
\arguments{
\item{sym_name}{Vector or List. A sequence of character symbol names.}

\item{sym_value}{Vector or List. A sequence of symbol values.}
}
\value{
Return the object itself.
}
\description{
This function stores the user inputs as the updated values of
symbols. List of symbol names and symbol types will be updated automatically.
There is no protection for values that should not be modified, which may lead
to error or loss of binding of some objects. Please use this function with
caution.
}
\examples{

# Constant variable
a <- 1

# Random uniform variable
z <- closed_form(~2 + a)

z$compute()

# Update value for "a"
z$set_sym("a", 3)

z$compute()

# Update value for "+"
z$set_sym("+", list(function(a, b) a * b))

z$compute()
}
