% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM$sym_name}
\alias{CLOSED_FORM$sym_name}
\title{List of symbol names in the abstract syntax tree of the expression}
\description{
A list, will be initialized after an instance is built.
}
\seealso{
Examples and usages: \link{CLOSED_FORM$sym}
}
