% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\docType{data}
\name{CLOSED_FORM}
\alias{CLOSED_FORM}
\title{CLOSED_FORM class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
CLOSED_FORM
}
\description{
This is the base class of closed form expression, inherited from
\link{BASE}.
}
\seealso{
Parent class: \link{BASE}
\cr
\cr
New attributes: \link{CLOSED_FORM$sym}, \link{CLOSED_FORM$sym_name},
\link{CLOSED_FORM$sym_type}, \link{CLOSED_FORM$expr}
\cr
\cr
New methods: \link{CLOSED_FORM$..init..}, \link{CLOSED_FORM$..str..},
\link{CLOSED_FORM$..len..},
\link{CLOSED_FORM$compute}, \link{CLOSED_FORM$gen}, \link{CLOSED_FORM$ast},
\link{CLOSED_FORM$as_dataframe}
}
\keyword{datasets}
