% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{define_pkg_fn}
\alias{define_pkg_fn}
\title{Load functions from package namespaces into current environment}
\usage{
define_pkg_fn(pkg, ...)
}
\arguments{
\item{pkg}{Package.}

\item{...}{Functions. Preferred names can be provide via named arguments.}
}
\value{
No return value, called for side effects.
}
\description{
This function loads functions from package namespaces and
assigns them to the preferred function names in the current environment.
}
\details{
Preferred function names can be provide via named arguments
like \code{info = cli_alert_info}.
}
\examples{
define_pkg_fn(pkg = cli, cli_alert_info, cli_alert_warning)
define_pkg_fn(cli, cli_alert_warning, info = cli_alert_info)

}
