% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%-=\%}
\alias{\%-=\%}
\title{In-place subtraction operator}
\usage{
x \%-=\% y
}
\arguments{
\item{x}{Object.}

\item{y}{Another object.}
}
\value{
Depends on the method.
}
\description{
In-place subtraction operator
}
\examples{
COMPANY <- new_class(class_name = "COMPANY")
company <- COMPANY$instantiate
register_method(COMPANY,
                ..init.. = function(name, age) {
                  self$name <- name
                  self$age <- age
                })
register_method(COMPANY,
                ..isub.. = function(y) {
                  self$age <- self$age[self$name != y]
                  self$name <- self$name[self$name != y]
                })

good_company <- company(c("patrick", "james"),
                        c(33, 34))
good_company \%-=\% "patrick"
good_company$name

}
