% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%gt\%.bandicoot_oop}
\alias{\%gt\%.bandicoot_oop}
\title{S3 method of performing the greater than operator of a
\code{bandicoot_oop} object}
\usage{
\method{\%gt\%}{bandicoot_oop}(x, y)
}
\arguments{
\item{x}{\code{bandicoot_oop} object.}

\item{y}{Object.}
}
\value{
A Boolean value.
}
\description{
This function performs the greater than operator using the \code{..gt..()} method.
If it is not applicable, error will be raised.
}
