% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%gt\%}
\alias{\%gt\%}
\title{The greater than operator}
\usage{
x \%gt\% y
}
\arguments{
\item{x}{Object.}

\item{y}{Another object.}
}
\value{
A Boolean value.
}
\description{
The greater than operator
}
\examples{
NAME <- new_class(class_name = "NAME")
name <- NAME$instantiate
register_method(NAME,
                ..init.. = function(first_name, last_name) {
                  self$first_name = first_name
                  self$last_name = last_name
                })
register_method(NAME,
                ..gt.. = function(y) {
                  if (self$last_name == y$last_name) {
                    return(self$first_name > y$first_name)
                  }
                  return(self$last_name > self$last_name)
                })

patrick <- name("Patrick", "Li")
james <- name("James", "Li")
patrick \%gt\% james

}
