% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%le\%}
\alias{\%le\%}
\title{The less or equals to operator}
\usage{
x \%le\% y
}
\arguments{
\item{x}{Object.}

\item{y}{Another object.}
}
\value{
A Boolean value.
}
\description{
The less or equals to operator
}
