% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%ne\%}
\alias{\%ne\%}
\title{The not equals to operator}
\usage{
x \%ne\% y
}
\arguments{
\item{x}{Object.}

\item{y}{Another object.}
}
\value{
A Boolean value.
}
\description{
The not equals to operator
}
\examples{
AGE <- new_class(class_name = "AGE")
age <- AGE$instantiate
register_method(AGE, ..init.. = function(current) self$current = current)
register_method(AGE, ..ne.. = function(y) self$current != y$current)

patrick <- age(33)
james <- age(33)
patrick \%ne\% james

}
