% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%+\%}
\alias{\%+\%}
\title{Addition operator}
\usage{
x \%+\% y
}
\arguments{
\item{x}{Object.}

\item{y}{Another object.}
}
\value{
Depends on the method.
}
\description{
Addition operator
}
\examples{
COMPANY <- new_class(class_name = "COMPANY")
company <- COMPANY$instantiate
register_method(COMPANY,
                ..init.. = function(name, age) {
                  self$name <- name
                  self$age <- age
                })
register_method(COMPANY,
                ..add.. = function(y) {
                  company(c(self$name, y$name),
                          c(self$age, y$age))
                })

good_company <- company(c("patrick", "james"),
                        c(33, 34))
bad_company <- company(c("pat", "jam"),
                       c(3, 4))
new_company <- good_company \%+\% bad_company
new_company$name

}
