% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{is_bandicoot_oop}
\alias{is_bandicoot_oop}
\title{Check whether the object is a \code{bandicoot_oop} object}
\usage{
is_bandicoot_oop(obj, why = FALSE)
}
\arguments{
\item{obj}{Any object.}

\item{why}{Boolean. Whether or not to print the reason when the check fail.}
}
\value{
A Boolean value.
}
\description{
This function check whether the object is a \code{bandicoot_oop} object.
}
\examples{

e <- new.env()
is_bandicoot_oop(e)

e <- new_class(class_name = "test")
is_bandicoot_oop(e)

}
