% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{iter}
\alias{iter}
\title{Build an iterator}
\usage{
iter(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments needed for building an iterator.}
}
\value{
An iterator.
}
\description{
Build an iterator
}
