% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{len}
\alias{len}
\title{Compute the length of the object}
\usage{
len(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments needed for computing the length.}
}
\value{
An integer.
}
\description{
Compute the length of the object
}
\examples{
EMAIL <- new_class(class_name = "EMAIL")
email <- EMAIL$instantiate
register_method(EMAIL, ..init.. = function(address) self$address = address)
register_method(EMAIL, ..len.. = function() nchar(self$address))

patrick <- email('patrick@test.com')
len(patrick)

}
