% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_zzz.R
\name{portal}
\alias{portal}
\alias{base_}
\alias{rand_var}
\alias{rand_uniform}
\alias{rand_uniform_d}
\alias{rand_normal}
\alias{rand_lognormal}
\alias{closed_form}
\title{Portals to class instantiate methods}
\usage{
base_(..., env = new.env(parent = parent.frame()))

rand_var(..., env = new.env(parent = parent.frame()))

rand_uniform(..., env = new.env(parent = parent.frame()))

rand_uniform_d(..., env = new.env(parent = parent.frame()))

rand_normal(..., env = new.env(parent = parent.frame()))

rand_lognormal(..., env = new.env(parent = parent.frame()))

closed_form(..., env = new.env(parent = parent.frame()))
}
\arguments{
\item{...}{Arguments passed to \code{init} method.}

\item{env}{Environment. The instance environment.}
}
\description{
These functions are the same as \code{class$instantiate}.
}
\section{Functions}{
\itemize{
\item \code{base_}: \link{BASE}

\item \code{rand_var}: \link{RAND_VAR}

\item \code{rand_uniform}: \link{RAND_UNIFORM}

\item \code{rand_uniform_d}: \link{RAND_UNIFORM_D}

\item \code{rand_normal}: \link{RAND_NORMAL}

\item \code{rand_lognormal}: \link{RAND_LOGNORMAL}

\item \code{closed_form}: \link{CLOSED_FORM}
}}

