% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{print.bandicoot_oop}
\alias{print.bandicoot_oop}
\title{S3 method of printing \code{bandicoot_oop} object}
\usage{
\method{print}{bandicoot_oop}(x, ...)
}
\arguments{
\item{x}{\code{bandicoot_oop} object.}

\item{...}{ignored.}
}
\value{
No return value, called for side effects.
}
\description{
This function print the string representation of the object by using the
\code{..str..()} method.
}
\examples{
print(base_())
}
