% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_LOGNORMAL$..init..}
\alias{RAND_LOGNORMAL$..init..}
\title{Initialization method}
\arguments{
\item{mu}{Numeric. Mean of the log of the random variable. Default is 0.}

\item{sigma}{Numeric. Standard deviation of the log the random variable.
Default is 1.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
}
\examples{

# Instantiate
test <- rand_lognormal(mu = 1, sigma = 2)
test
}
