% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\docType{data}
\name{RAND_LOGNORMAL}
\alias{RAND_LOGNORMAL}
\title{RAND_LOGNORMAL class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
RAND_LOGNORMAL
}
\description{
This is the class of the log-normal random variable, inherited from
\link{RAND_VAR}.
}
\seealso{
Parent class: \link{RAND_VAR}
\cr
\cr
New methods: \link{RAND_LOGNORMAL$..init..}, \link{RAND_LOGNORMAL$gen},
RAND_LOGNORMAL$E, RAND_LOGNORMAL$Var
}
\keyword{datasets}
