% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_NORMAL$..init..}
\alias{RAND_NORMAL$..init..}
\title{Initialization method}
\arguments{
\item{mu}{Numeric. Mean. Default is 0.}

\item{sigma}{Numeric. Standard deviation. Default is 1.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
}
\examples{

# Instantiate
test <- rand_normal(mu = 1, sigma = 2)
test
}
