% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_NORMAL$gen}
\alias{RAND_NORMAL$gen}
\title{Generate random values}
\arguments{
\item{n}{Integer. Number of observations.}

\item{mu}{Numeric. If it is provided and has length \code{n}, values will be used in
each realization.}

\item{sigma}{Numeric. If it is provided and has length \code{n}, values will be used in
each realization.}
}
\value{
A vector of numeric values.
}
\description{
This function generates random values from the random variable.
}
\examples{

test <- rand_normal(mu = 1, sigma = 2)
test$gen(10)

test$gen(3, mu = c(0,1,2), sigma = c(1,2,4))
}
\seealso{
\code{\link[stats:Normal]{stats::rnorm()}}
}
