% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\docType{data}
\name{RAND_NORMAL}
\alias{RAND_NORMAL}
\title{RAND_NORMAL class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
RAND_NORMAL
}
\description{
This is the class of the normal random variable, inherited from
\link{RAND_VAR}.
}
\seealso{
Parent class: \link{RAND_VAR}
\cr
\cr
New methods: \link{RAND_NORMAL$..init..}, \link{RAND_NORMAL$gen},
RAND_NORMAL$E, RAND_NORMAL$Var
}
\keyword{datasets}
