% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_UNIFORM$gen}
\alias{RAND_UNIFORM$gen}
\title{Generate random values}
\arguments{
\item{n}{Integer. Number of observations.}

\item{a}{Numeric. If it is provided and has length \code{n}, values will be used in
each realization.}

\item{b}{Numeric. If it is provided and has length \code{n}, values will be used in
each realization.}
}
\value{
A vector of numeric values.
}
\description{
This function generates random values from the random variable.
}
\examples{

test <- rand_uniform(a = 1, b = 2)
test$gen(10)

test$gen(3, a = c(1,2,3), b = c(2,3,4))
}
\seealso{
\code{\link[stats:Uniform]{stats::runif()}}
}
