% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\docType{data}
\name{RAND_UNIFORM}
\alias{RAND_UNIFORM}
\title{RAND_UNIFORM class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
RAND_UNIFORM
}
\description{
This is the class of the uniform random variable, inherited from
\link{RAND_VAR}.
}
\seealso{
Parent class: \link{RAND_VAR}
\cr
\cr
New methods: \link{RAND_UNIFORM$..init..}, \link{RAND_UNIFORM$gen},
RAND_UNIFORM$E, RAND_UNIFORM$Var
}
\keyword{datasets}
