% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_UNIFORM_D$..init..}
\alias{RAND_UNIFORM_D$..init..}
\title{Initialization method}
\arguments{
\item{a}{Numeric. Lower bound. Default is 0.}

\item{b}{Numeric. Upper bound. Default is 1.}

\item{k}{Integer. Number of unique discrete values. Default is 5.}

\item{even}{Boolean. Whether or not candidate values are evenly spaced.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
}
\examples{

# Instantiate
test <- rand_uniform_d(a = 1, b = 2, k = 3)
test
}
