% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_UNIFORM_D$gen}
\alias{RAND_UNIFORM_D$gen}
\title{Generate random values}
\arguments{
\item{n}{Integer. Number of observations.}

\item{a}{Numeric. If it is provided and has length \code{n}, values will be used in
each realization.}

\item{b}{Numeric. If it is provided and has length \code{n}, values will be used in
each realization.}

\item{k}{Numeric. If it is provided and has length \code{n}, values will be used in
each realization.}

\item{even}{Boolean. If it is provided and has length \code{n}, values will be used in
each realization.}
}
\value{
A vector of numeric values.
}
\description{
This function generates random values from the random variable.
}
\examples{

test <- rand_uniform_d(a = 1, b = 2, k = 2, even = TRUE)
test$gen(10)

test$gen(3, a = c(1,2,3), b = c(2,3,4), k = 1, even = c(TRUE, TRUE, FALSE))
}
\seealso{
\code{\link[stats:Uniform]{stats::runif()}}, \code{\link[=sample]{sample()}}
}
