% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\docType{data}
\name{RAND_UNIFORM_D}
\alias{RAND_UNIFORM_D}
\title{RAND_UNIFORM_D class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
RAND_UNIFORM_D
}
\description{
This is the class of the discrete uniform random variable, inherited from
\link{RAND_VAR}.
}
\seealso{
Parent class: \link{RAND_VAR}
\cr
\cr
New methods: \link{RAND_UNIFORM_D$..init..}, \link{RAND_UNIFORM_D$gen},
RAND_UNIFORM_D$E, RAND_UNIFORM_D$Var
}
\keyword{datasets}
