% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_VAR$..init..}
\alias{RAND_VAR$..init..}
\title{Initialization method}
\arguments{
\item{dist}{Character. Distribution name.}

\item{prm}{List. List of parameters.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
}
\examples{

RAND_VAR$..init..

# Instantiate
test <- rand_var(dist = "uniform", prm = list(a = 1, b = 2))
test$prm
test$dist
}
