% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_VAR$..str..}
\alias{RAND_VAR$..str..}
\title{String representation of the object}
\value{
A string.
}
\description{
This function returns a string representation of the object.
}
\examples{

RAND_VAR$..str..()

test <- rand_var(dist = "uniform", prm = list(a = 1, b = 2))
test$..str..()
}
