% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_VAR$E}
\alias{RAND_VAR$E}
\title{Expectation of the random variable}
\value{
NA
}
\description{
Expectation of the random variable.
User could override this method in derived class.
}
\examples{

RAND_VAR$E()
}
