% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_VAR$prm}
\alias{RAND_VAR$prm}
\title{List of parameters}
\description{
A list, will be initialized after an instance is built.
}
\examples{

# Instantiate
test <- rand_var(dist = "uniform", prm = list(a = 1, b = 2))
test$prm
test$dist
}
