% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_VAR$set_prm}
\alias{RAND_VAR$set_prm}
\title{Generate random values}
\arguments{
\item{prm_name}{List or Vector. A sequence of character parameter names.}

\item{prm_value}{List or Vector. A sequence of parameter values.}
}
\value{
Return the object itself.
}
\description{
This function updates the parameters.
}
\examples{

test <- rand_var()
test$set_prm("a", 1)
test
}
