% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{repr.bandicoot_oop}
\alias{repr.bandicoot_oop}
\title{S3 method of computing the "official" string representation of a
\code{bandicoot_oop} object}
\usage{
\method{repr}{bandicoot_oop}(x, ...)
}
\arguments{
\item{x}{\code{bandicoot_oop} object.}

\item{...}{ignored.}
}
\value{
An integer.
}
\description{
This function computes the "official" string representation of a
\code{bandicoot_oop} object using the \code{..repr..()} method.
If it is not applicable, error will be raised.
}
\examples{
EMAIL <- new_class(class_name = "EMAIL")
email <- EMAIL$instantiate
register_method(EMAIL, ..init.. = function(address) self$address = address)

patrick <- email('patrick@test.com')
repr(patrick)

}
