% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{repr}
\alias{repr}
\title{The "official" string representation of an object.}
\usage{
repr(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments needed for computing the string.}
}
\value{
A string.
}
\description{
The "official" string representation of an object. If at all possible,
this should look like a valid R expression that could be used to recreate
an object wit the same value (given an appropriate environment). This
description is copied from the python documentation.
}
