% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..len..}
\alias{BASE$..len..}
\title{Length of the class or the instance}
\description{
User could override this method in derived class.
}
\examples{

BASE$..len..()

# Inherit from BASE
TEST <- new_class(BASE, class_name = "TEST")

# Override the `..len..` method
register_method(TEST, ..len.. = function() 1)
TEST$..len..()
}
