% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$..methods..}
\alias{BASE$..methods..}
\title{List all methods of a class or an instance}
\value{
A string vector.
}
\description{
This function lists all methods of a class or an instance
}
\examples{

BASE$..methods..()
}
