% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\name{BASE$set_attr}
\alias{BASE$set_attr}
\title{Set value of an attribute or a method}
\arguments{
\item{attr_name}{Character. Attribute name.}

\item{attr_val}{Any value.}
}
\value{
Return the object itself.
}
\description{
This function sets the value of an attribute or a method.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{BASE$set_attr(attr_name, attr_val)
}\if{html}{\out{</div>}}
}
}
\examples{

test <- BASE$instantiate()
test$set_attr("x", 1)
test$x
}
