% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\docType{data}
\name{BASE}
\alias{BASE}
\title{BASE class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
BASE
}
\description{
This class provides essential attributes and methods. It makes
the assumption that the container name is \code{..method_env..} and the name of
the reference to self is \code{self}. If you would like to use other container
names and self names, you need to overwrite the class definition of BASE.
}
\seealso{
Attributes: \link{BASE$..type..}, \link{BASE$..class..}, \link{BASE$..method_env..},
\link{BASE$..instantiated..}
\cr
\cr
Methods: \link{BASE$..dict..}, \link{BASE$..str..}, \link{BASE$..len..},
\link{BASE$has_attr}, \link{BASE$get_attr}, \link{BASE$set_attr}, \link{BASE$del_attr},
\link{BASE$..methods..}, \link{BASE$..init..}, \link{BASE$..new..}, \link{BASE$instantiate}
}
\keyword{datasets}
