% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%contains\%}
\alias{\%contains\%}
\title{Membership test operator}
\usage{
x \%contains\% y
}
\arguments{
\item{x}{Object.}

\item{y}{Another object.}
}
\value{
A Boolean value.
}
\description{
Membership test operator
}
