% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%-=\%}
\alias{\%-=\%}
\title{In-place subtraction operator}
\usage{
x \%-=\% y
}
\arguments{
\item{x}{Object.}

\item{y}{Another object.}
}
\value{
Depends on the method.
}
\description{
In-place subtraction operator
}
