% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%ge\%.bandicoot_oop}
\alias{\%ge\%.bandicoot_oop}
\title{S3 method of performing the greater or equals operator of a
\code{bandicoot_oop} object}
\usage{
\method{\%ge\%}{bandicoot_oop}(x, y)
}
\arguments{
\item{x}{\code{bandicoot_oop} object.}

\item{y}{Object.}
}
\value{
A Boolean value.
}
\description{
This function performs the greater or equals operator using the \code{..ge..()} method.
If it is not applicable, error will be raised.
}
