% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%ge\%}
\alias{\%ge\%}
\title{The greater or equals to operator}
\usage{
x \%ge\% y
}
\arguments{
\item{x}{Object.}

\item{y}{Another object.}
}
\value{
A Boolean value.
}
\description{
The greater or equals to operator
}
