% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{\%lt\%.bandicoot_oop}
\alias{\%lt\%.bandicoot_oop}
\title{S3 method of performing the less than operator of a
\code{bandicoot_oop} object}
\usage{
\method{\%lt\%}{bandicoot_oop}(x, y)
}
\arguments{
\item{x}{\code{bandicoot_oop} object.}

\item{y}{Object.}
}
\value{
A Boolean value.
}
\description{
This function performs the less than operator using the \code{..lt..()} method.
If it is not applicable, error will be raised.
}
