% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{import_bandicoot}
\alias{import_bandicoot}
\title{Load functions from the bandicoot into target environment or search path}
\usage{
import_bandicoot(
  env = parent.frame(),
  package = FALSE,
  reload = FALSE,
  import_oop = TRUE,
  import_base = TRUE,
  import_rand_var = FALSE,
  import_closed_form = FALSE
)
}
\arguments{
\item{env}{Environment. The target environment.}

\item{package}{Boolean. Whether or not it is used in a package.}

\item{reload}{Boolean. Whether or not to reload the namespace. Only works if
\code{package = FALSE}.}

\item{import_oop}{Boolean. Whether or not to import OOP tools.}

\item{import_base}{Boolean. Whether or not to import BASE class.}

\item{import_rand_var}{Boolean. Whether or not to import RAND_VAR classes.}

\item{import_closed_form}{Boolean. Whether or not to import CLOSED_FORM class.}
}
\value{
No return value, called for side effects.
}
\description{
This function is critical when other packages want to use the bandicoot OOP
system. Since this OOP system is based on environment, any instance will only
run on the environment they defined. So, function like \link{use_method} which
only exists in the package environment can not be accessed by the instance,
unless the function has been loaded into the current environment by calling
\code{use_method <- bandicoot::use_method}, or \code{require(bandicoot)} or
\code{library(bandicoot)}. This issue can usually be addressed by
using the package name directly inside the method body like
\code{this_method <- function() bandicoot::use_method()}. However, if it is not
possible, then this function helps loads corresponding function into target
environment or search path.
}
\details{
If it is used in a package, specify \code{package = TRUE}, this function will
call \link{define_pkg_fn}. Otherwise, it will call \link{require} to attach the
functions into the search path.
}
