% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{iter.bandicoot_oop}
\alias{iter.bandicoot_oop}
\title{S3 method of building an iterator of a
\code{bandicoot_oop} object}
\usage{
\method{iter}{bandicoot_oop}(x, ...)
}
\arguments{
\item{x}{\code{bandicoot_oop} object.}

\item{...}{Additional arguments needed for building an iterator.}
}
\value{
An iterator.
}
\description{
This function builds an iterator using the \code{..iter..()} method.
If it is not applicable, error will be raised.
}
