% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{len.bandicoot_oop}
\alias{len.bandicoot_oop}
\title{S3 method of computing the length of \code{bandicoot_oop} object}
\usage{
\method{len}{bandicoot_oop}(x, ...)
}
\arguments{
\item{x}{\code{bandicoot_oop} object.}

\item{...}{ignored.}
}
\value{
An integer.
}
\description{
This function computes the length of the object by using the \code{..len..()}
method. If it is not applicable, error will be raised.
}
