% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_s3.R
\name{len}
\alias{len}
\title{Compute the length of the object}
\usage{
len(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments needed for computing the length.}
}
\value{
An integer.
}
\description{
Compute the length of the object
}
