% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_zzz.R
\name{portal}
\alias{portal}
\alias{base_}
\title{Portals to class instantiate methods}
\usage{
base_(..., env = new.env(parent = parent.frame()))
}
\arguments{
\item{...}{Arguments passed to \code{init} method.}

\item{env}{Environment. The instance environment.}
}
\description{
These functions are the same as \code{class$instantiate}.
}
\section{Functions}{
\itemize{
\item \code{base_}: \link{BASE}
}}

